refactor: 분리된 컴포넌트 Profile과 MenuContainer를 대시보드에 포함
