--- src/pages/dashboard/components/MenuContainer/index.ts
@@ -1,3 +1,0 @@
-import MenuContainer from './MenuContainer';
-
-export { MenuContainer };
