--- src/pages/dashboard/components/Profile/index.ts
@@ -1,3 +1,0 @@
-import Profile from './Profile';
-
-export { Profile };
