+++ src/pages/dashboard/components/index.ts
@@ -1,15 +1,6 @@
-import { MenuContainer } from './MenuContainer';
-import { Profile } from './Profile';
 import { NotSelected } from './NotSelected';
 import { MemberDetail } from './MemberDetail';
 import { PrescriptionManagement } from './PrescriptionManagement';
 import { MemberManagement } from './MemberManagement';
 
-export {
-  MenuContainer,
-  Profile,
-  NotSelected,
-  MemberDetail,
-  PrescriptionManagement,
-  MemberManagement,
-};
+export { NotSelected, MemberDetail, PrescriptionManagement, MemberManagement };
